# app/Http/Requests/GenreRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;


/**
 * @OA\Schema(
 *  schema="GenreRequest",
 *  title="GenreRequest",
 * 	@OA\Property(
 * 		property="name",
 * 		type="string"
 * 	)
 * )
 */
class GenreRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     */
    public function rules(): array
    {
        return [
            'name' => 'required|string|max:255',
        ];
    }
}