# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Libreria",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/genres": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Get all genres",
                "description": "Retrieve a list of all genres, cached for 5 minutes.",
                "operationId": "eb12635064fefadf7ff7e54900a73534",
                "responses": {
                    "200": {
                        "description": "A list of genres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Genre"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Genres"
                ],
                "summary": "Create a new genre",
                "description": "Add a new genre to the database.",
                "operationId": "8596dc3b1780529a7ce51429eee5ed0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Genre created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/genres/{id}": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Get genre details",
                "description": "Retrieve the details of a specific genre by ID.",
                "operationId": "3e2ac2805d9f25d984bab075133bbb96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the genre to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Genre details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Genre not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Genres"
                ],
                "summary": "Update a genre",
                "description": "Update the details of an existing genre by ID.",
                "operationId": "65d9a201f51320eb4147a75f91a8119b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the genre to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Genre updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Genre not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Genres"
                ],
                "summary": "Delete a genre",
                "description": "Delete a specific genre by ID.",
                "operationId": "4ddc050edb0bf18c202c611a028f457f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the genre to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Genre deleted successfully"
                    },
                    "404": {
                        "description": "Genre not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "GenreRequest": {
                "title": "GenreRequest",
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Genre": {
                "title": "Genre",
                "description": "Genre model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Fantasy"
                    },
                    "slug": {
                        "type": "string",
                        "example": "fantasy"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T01:00:00Z"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}